+++ core/scs_profile/__init__.py
@@ -2,6 +2,8 @@
 
 from .intensity import (
     BrakeIntensityCache,
+    LowBrakeIntensityAebWarning,
+    aeb_max_brake_ms2,
     apply_brake_intensity,
     learn_decel_scale,
 )
@@ -9,7 +11,9 @@ from .reader import SelectedProfileSettings, brake_ui_scale, read_selected_profi
 
 __all__ = [
     "BrakeIntensityCache",
+    "LowBrakeIntensityAebWarning",
     "SelectedProfileSettings",
+    "aeb_max_brake_ms2",
     "apply_brake_intensity",
     "brake_ui_scale",
     "learn_decel_scale",